move LoginPage.css and apply to components
