+++ acm_website/src/pages/LoginPage.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import './LoginPage.css'; // Import the CSS file for styling
+import './../styles/LoginPage.css';
 import { auth } from '../firebase/config';
 import { createUserWithEmailAndPassword,
          onAuthStateChanged,
